QOL update
Features:
Now experience label is a button, and when you click on it, it will level up the corresponding identity.

Fixes:
Working around a bug, when failed notches cannot be reset to zero. Quick and lazy fix, but it works.

Localizations:
Added Russian localization.
Changed the CSS table a little. In Russian some text is wider. Needed more space for it.